+++ test/Nerdbank.MessagePack.TUnit/SerializationCallbackTests.cs
@@ -8,16 +8,16 @@ public partial class SerializationCallbackTests : MessagePackSerializerTestBase
 	private readonly BusyClassArray objectAsArray = new() { Name = "Alice" };
 	private readonly BusyClassMap objectAsMap = new() { Name = "Alice" };
 
-	[Theory, PairwiseData]
+	[Test, MatrixDataSource]
 	public async Task Serialize_ObjectAsArray(bool async)
 	{
 		if (async)
 		{
-			await this.Serializer.SerializeAsync(Stream.Null, this.objectAsArray, TestContext.Current.CancellationToken);
+			await this.Serializer.SerializeAsync(Stream.Null, this.objectAsArray, this.TimeoutToken);
 		}
 		else
 		{
-			this.Serializer.Serialize(this.objectAsArray, TestContext.Current.CancellationToken);
+			this.Serializer.Serialize(this.objectAsArray, this.TimeoutToken);
 		}
 
 		Assert.Equal(1, this.objectAsArray.OnBeforeSerializeCounter);
@@ -26,16 +26,16 @@ public partial class SerializationCallbackTests : MessagePackSerializerTestBase
 		Assert.Equal(0, this.objectAsArray.OnAfterDeserializeCounter);
 	}
 
-	[Theory, PairwiseData]
+	[Test, MatrixDataSource]
 	public async Task Serialize_ObjectAsMap(bool async)
 	{
 		if (async)
 		{
-			await this.Serializer.SerializeAsync(Stream.Null, this.objectAsMap, TestContext.Current.CancellationToken);
+			await this.Serializer.SerializeAsync(Stream.Null, this.objectAsMap, this.TimeoutToken);
 		}
 		else
 		{
-			this.Serializer.Serialize(this.objectAsMap, TestContext.Current.CancellationToken);
+			this.Serializer.Serialize(this.objectAsMap, this.TimeoutToken);
 		}
 
 		Assert.Equal(1, this.objectAsMap.OnBeforeSerializeCounter);
@@ -44,12 +44,12 @@ public partial class SerializationCallbackTests : MessagePackSerializerTestBase
 		Assert.Equal(0, this.objectAsMap.OnAfterDeserializeCounter);
 	}
 
-	[Theory, PairwiseData]
+	[Test, MatrixDataSource]
 	public async Task Deserialize_ObjectAsArray(bool async)
 	{
 		BusyClassArray? obj = async
-			? await this.Serializer.DeserializeAsync<BusyClassArray>(PipeReader.Create(new(ObjectAsArrayMsgPack)), TestContext.Current.CancellationToken)
-			: this.Serializer.Deserialize<BusyClassArray>(ObjectAsArrayMsgPack, TestContext.Current.CancellationToken);
+			? await this.Serializer.DeserializeAsync<BusyClassArray>(PipeReader.Create(new(ObjectAsArrayMsgPack)), this.TimeoutToken)
+			: this.Serializer.Deserialize<BusyClassArray>(ObjectAsArrayMsgPack, this.TimeoutToken);
 		Assert.NotNull(obj);
 
 		Assert.Equal(0, obj.OnBeforeSerializeCounter);
@@ -58,12 +58,12 @@ public partial class SerializationCallbackTests : MessagePackSerializerTestBase
 		Assert.Equal(1, obj.OnAfterDeserializeCounter);
 	}
 
-	[Theory, PairwiseData]
+	[Test, MatrixDataSource]
 	public async Task Deserialize_ObjectAsMap(bool async)
 	{
 		BusyClassMap? obj = async
-			? await this.Serializer.DeserializeAsync<BusyClassMap>(PipeReader.Create(new(ObjectAsMapMsgPack)), TestContext.Current.CancellationToken)
-			: this.Serializer.Deserialize<BusyClassMap>(ObjectAsMapMsgPack, TestContext.Current.CancellationToken);
+			? await this.Serializer.DeserializeAsync<BusyClassMap>(PipeReader.Create(new(ObjectAsMapMsgPack)), this.TimeoutToken)
+			: this.Serializer.Deserialize<BusyClassMap>(ObjectAsMapMsgPack, this.TimeoutToken);
 		Assert.NotNull(obj);
 
 		Assert.Equal(0, obj.OnBeforeSerializeCounter);
@@ -72,12 +72,12 @@ public partial class SerializationCallbackTests : MessagePackSerializerTestBase
 		Assert.Equal(1, obj.OnAfterDeserializeCounter);
 	}
 
-	[Theory, PairwiseData]
+	[Test, MatrixDataSource]
 	public async Task Deserialize_ObjectAsArray_Init(bool async)
 	{
 		BusyClassArray? obj = async
-			? await this.Serializer.DeserializeAsync<BusyClassArrayInit>(PipeReader.Create(new(ObjectAsArrayMsgPack)), TestContext.Current.CancellationToken)
-			: this.Serializer.Deserialize<BusyClassArrayInit>(ObjectAsArrayMsgPack, TestContext.Current.CancellationToken);
+			? await this.Serializer.DeserializeAsync<BusyClassArrayInit>(PipeReader.Create(new(ObjectAsArrayMsgPack)), this.TimeoutToken)
+			: this.Serializer.Deserialize<BusyClassArrayInit>(ObjectAsArrayMsgPack, this.TimeoutToken);
 		Assert.NotNull(obj);
 
 		Assert.Equal(0, obj.OnBeforeSerializeCounter);
@@ -86,12 +86,12 @@ public partial class SerializationCallbackTests : MessagePackSerializerTestBase
 		Assert.Equal(1, obj.OnAfterDeserializeCounter);
 	}
 
-	[Theory, PairwiseData]
+	[Test, MatrixDataSource]
 	public async Task Deserialize_ObjectAsMap_Init(bool async)
 	{
 		BusyClassMap? obj = async
-			? await this.Serializer.DeserializeAsync<BusyClassMapInit>(PipeReader.Create(new(ObjectAsMapMsgPack)), TestContext.Current.CancellationToken)
-			: this.Serializer.Deserialize<BusyClassMapInit>(ObjectAsMapMsgPack, TestContext.Current.CancellationToken);
+			? await this.Serializer.DeserializeAsync<BusyClassMapInit>(PipeReader.Create(new(ObjectAsMapMsgPack)), this.TimeoutToken)
+			: this.Serializer.Deserialize<BusyClassMapInit>(ObjectAsMapMsgPack, this.TimeoutToken);
 		Assert.NotNull(obj);
 
 		Assert.Equal(0, obj.OnBeforeSerializeCounter);
